Updated Canvas.
Split the .py file to show splash screen earlier, during loading.
